DERBY-6000: Add new JDBC 4.2 Statement.executeLargeUpdate() methods to the embeddeded JDBC 3.0 implementation.
git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1438600 13f79535-47bb-0310-9956-ffa450edef68

+++ java/engine/org/apache/derby/iapi/sql/ResultSet.java
@@ -83,7 +83,7 @@ public interface ResultSet
 	 *
 	 * @return	The number of rows affect by the statement, so far.
 	 */
-	int	modifiedRowCount();
+	long	modifiedRowCount();
 
 	/**
 	 * Returns a ResultDescription object, which describes the results
