+++ java/engine/org/apache/derby/impl/sql/execute/NoRowsResultSetImpl.java
@@ -127,7 +127,7 @@ abstract class NoRowsResultSetImpl implements ResultSet
 	/**
 	 * Returns zero.
 	 */
-	public int	modifiedRowCount() { return 0; }
+	public long	modifiedRowCount() { return 0L; }
 
 	/**
 	 * Returns null.
